added component files and refactored existing files to facilitate components. All working as expected
