+++ todo-list/src/AddTodoForm.js
@@ -1,0 +1,11 @@
+import React from 'react';
+
+const AddTodoForm = () => (
+    <form>
+        <label htmlFor="TodoTitle"> Title </label>
+        <input id="TodoTitle"></input>
+        <button type="submit"> Add </button>
+    </form>
+)
+
+export default AddTodoForm;
